rake task to get usd rate and model update

+++ db/migrate/20191109170026_create_usd_exchange_rates.rb
@@ -1,7 +1,7 @@
 class CreateUsdExchangeRates < ActiveRecord::Migration[5.2]
   def change
     create_table :usd_exchange_rates do |t|
-      t.integer  :rate, null: false
+      t.float    :rate, null: false
       t.boolean  :is_forced
       t.datetime :expiration_date
       t.timestamps
